requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asyncpg==0.30.0
azure-common==1.1.28
azure-core==1.34.0
azure-functions==1.23.0
azure-identity==1.23.0
azure-keyvault-secrets==4.10.0
azure-search-documents==11.5.3
azure-storage-blob==12.25.1
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
colorama==0.4.6
cryptography==45.0.4
dacite==1.9.2
distro==1.9.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
isodate==0.7.2
Jinja2==3.1.6
jiter==0.10.0
MarkupSafe==3.0.2
msal==1.32.3
msal-extensions==1.3.1
openai==1.91.0
packaging==25.0
pluggy==1.6.0
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.10.1
pyodbc==5.2.0
pytest==8.4.1
PyYAML==6.0.2
-e c:\repos\rag_framework
regex==2024.11.6
requests==2.32.4
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
tiktoken==0.9.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.5.0
Werkzeug==3.1.3
